requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
sqlparse==0.5.2
tzdata==2024.2
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
chistesESP==0.0.2
click==8.1.7
cohere==5.11.4
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
fastavro==1.9.7
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
MouseInfo==0.1.3
packaging==24.2
parameterized==0.9.0
pillow==11.0.0
PyAutoGUI==0.9.54
pydantic==2.10.1
pydantic_core==2.27.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywhatkit==5.4
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
tokenizers==0.20.3
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wikipedia==1.4.0
youtube-search-python==1.6.6
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
chistesESP==0.0.2
click==8.1.7
cohere==5.11.4
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
fastavro==1.9.7
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
MouseInfo==0.1.3
packaging==24.2
parameterized==0.9.0
pillow==11.0.0
PyAutoGUI==0.9.54
pydantic==2.10.1
pydantic_core==2.27.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pywhatkit==5.4
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
tokenizers==0.20.3
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wikipedia==1.4.0
youtube-search-python==1.6.6
